Added a version check for XML::DOM.
Improved compatibility for HTML escapes.

Update issue #5
Status: Fixed
The issue was fixed in the latest version of XML::DOM, 1.44, which was released in 2005.  Previous versions over used the 'use bytes' pragma.  Ubuntu 8.04 ships with version 1.43 released in 2003.  It might be time to find a better XML library for Perl.