Added parsing of ISO 8601 time stamps and options to gpx.pl.
New review
Summary: Verify time stamps are handled correctly from all sources.
Owner: penguin359
Time stamps are read in from various sources and formats.  Some are in local time and others are in UTC or at least have an offset from UTC.  Local time is probably best dealt by using the local time zone of the machine it is running on instead of a fixed offset that is used in certain places.  HG: Enter commit message.  Lines beginning with 'HG:' are removed.