Fixed bugs in OAuth client and improved debuging.
A recent bug was introduced which broke certain OAuth services.

A new CGI parameter, debug, was introduced to enabling debugging across the various steps of OAuth.  Most importantly, all redirects are printed out to the screen for manual execution.  Also, a CGI parameter, refresh, was introduced to start OAuth from scratch regardless of existing tokens.

Improved detection of step 4 in Oauth 1.0/1.0a.  Once the final step in OAuth/FlickrAuth is completed, redirect the user back to the original URL dropping any OAuth tokens passed so the page doesn't break if it's reloaded triggering the final step to restart incorrectly.